Safari-609.1.11.1/LayoutTests/fast/css3-text/css3-word-spacing-percentage/word-spacing-percentage-parse-expected.txt

PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('word-spacing') is "100%"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 1
PASS declaration.getPropertyValue('word-spacing') is "100.5%"
PASS cssRule.type is cssRule.STYLE_RULE
PASS declaration.length is 0
